% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshr.r
\name{refreshr}
\alias{refreshr}
\title{Package 'refreshr'}
\description{
Create refreshable dataframes/tables that automatically pull in
 data from an (internet) data source and transform the data (if neccessary)
 so that the user of your dataset does not have to worry about where to get
 the data from and how to update it.

 Functions available: \itemize{ \item \code{\link{make_refreshable}()}: Makes
 a dataframe/table refreshable. \item \code{\link{refresh}()}: Refreshes a
 dataframe/table. \item \code{\link{is.refreshr}()}: Checks if a
 dataframe/table is set up as refreshable. \item \code{\link{uptodate}()}:
 Checks if a refreshable dataframe/table is up to date compared to the remote
 data source. \item \code{\link{properties}()}: Prints or returns the main
 properties of a refreshable dataframe/table. }
}
